python/django/boilerplate/apps/core/exceptions.py
"""
Custom exceptions for the application.
"""
from rest_framework.views import exception_handler
from rest_framework.response import Response


def custom_exception_handler(exc, context):
    """
    Custom exception handler for API.
    Provides consistent error response format.
    """
    response = exception_handler(exc, context)
    
    if response is not None:
        custom_response_data = {
            'error': {
                'status_code': response.status_code,
                'message': response.data.get('detail', 'An error occurred'),
                'data': response.data
            }
        }
        response.data = custom_response_data
    
    return response

