python/django/boilerplate/project_name/settings/__init__.py
"""
Django settings module.
Import the appropriate settings based on environment.
"""
from decouple import config

ENVIRONMENT = config('ENVIRONMENT', default='development')

if ENVIRONMENT == 'production':
    from .production import *
elif ENVIRONMENT == 'testing':
    from .testing import *
else:
    from .development import *

